requirements.txt
anthropic>=0.49.0
streamlit>=1.40.0
chromadb>=0.5.0
pypdf>=4.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
